+++ src/main/java/onboarding/Problem2.java
@@ -1,6 +1,6 @@
 package onboarding;
 
-import onboarding.problem2.CryptogramService;
+import onboarding.problem2.service.CryptogramService;
 
 public class Problem2 {
     public static String solution(String cryptogram) {
