+++ src/test/java/onboarding/ApplicationTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
 
-import onboarding.problem2.CryptogramValidator;
+import onboarding.problem2.validation.CryptogramValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
